fix(services): remove dead LOCALE_REWRITE_PATHS, share PLAN_LABELS, validate signup plan
Three real issues surfaced when tracing every link on /en/services:

1. next.config.ts LOCALE_REWRITE_PATHS is dead code.

   next-intl middleware (localePrefix: "always") intercepts /services,

   /store, /contact, /blog, /docs, /case-studies, /work on its own and

   issues a 307 BEFORE Next config rewrites run. Verified live: both

   with and without the NEXT_LOCALE cookie a 307 still fires. The block

   was misleading documentation of a perf optimisation that doesn't

   exist. Deleted; /manifest.webmanifest -> /api/pwa-manifest is the

   only legit rewrite that remains.

2. PLAN_LABELS was duplicated in src/lib/pending-clients.ts (server)

   and src/app/portal/waiting/WaitingRoomClient.tsx (client) — drift

   risk. Extracted to a new dependency-free module src/lib/plans.ts

   that both can import (the client can't pull pending-clients.ts

   directly because it loads AWS SDK). Also exports isValidPlan() and

   PLAN_KEYS for reuse.

3. /auth/signup forwarded ?plan=<arbitrary-string> straight through to

   /portal/waiting, where the server enroll endpoint silently refused

   unknown plans — UX dead-end. Signup now validates ?plan= against the

   shared allowlist via isValidPlan() and drops bad values, so users

   land on /admin or /dashboard instead of a confused waiting room.

typecheck + lint + format all clean.
